refactor(ui): move auth logic to shared from features
- 인증 관련은 entities에서도 액세스가 필요하여, 레이어를 변경합니다.
