--- packages/ui/src/features/auth/index.ts
@@ -1,1 +1,0 @@
-export { login, logout } from './model/auth';
